grupp_sista (kopia)/requirements.txt
newspaper3k==0.2.8
pandas==1.5.3
numpy==1.24.3
transformers==4.30.2
torch==2.0.1
streamlit==1.24.0
scikit-learn==1.3.0
tqdm==4.65.0
sentencepiece==0.1.99 